requirements.txt
amqp==2.3.1
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
dnspython==1.15.0
fake-useragent==0.1.10
idna==2.6
kombu==4.2.0
PyMySQL==0.8.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
SQLAlchemy==1.2.8
urllib3==1.22
vine==1.1.4
